// packages/cc/src/lib/_Types.d.ts
export declare enum HumidityControlSetpointCommand {
    Set = 1,
    Get = 2,
    Report = 3,
    SupportedGet = 4,
    SupportedReport = 5,
    ScaleSupportedGet = 6,
    ScaleSupportedReport = 7,
    CapabilitiesGet = 8,
    CapabilitiesReport = 9
}