// packages/cc/src/lib/traits.d.ts
export interface CCParsingContext extends Readonly<SecurityManagers>, GetDeviceConfig, HostIDs {
    sourceNodeId: number;
    __internalIsMockNode?: boolean;
    /** If known, the frame type of the containing message */
    frameType: FrameType;
    getHighestSecurityClass(nodeId: number): MaybeNotKnown<SecurityClass>;
    hasSecurityClass(nodeId: number, securityClass: SecurityClass): MaybeNotKnown<boolean>;
    setSecurityClass(nodeId: number, securityClass: SecurityClass, granted: boolean): void;
}
/** Additional context needed for serializing CCs */
export interface CCEncodingContext extends Readonly<SecurityManagers>, GetDeviceConfig, HostIDs, GetSupportedCCVersion {
    getHighestSecurityClass(nodeId: number): MaybeNotKnown<SecurityClass>;
    hasSecurityClass(nodeId: number, securityClass: SecurityClass): MaybeNotKnown<boolean>;
    setSecurityClass(nodeId: number, securityClass: SecurityClass, granted: boolean): void;
}